backend/requirements.txt
 Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
Pillow==11.1.0
python-dotenv==1.0.1